requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Django==3.1.1
idna==2.10
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.2
mysqlclient==2.0.1
numpy==1.19.2
panda==0.3.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
urllib3==1.25.10
